requirements/base.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.2
django-environ==0.13.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
httplib2==0.31.2
idna==3.11
Markdown==3.10.2
oauthlib==3.3.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyparsing==3.3.2
redis==7.3.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.5
urllib3==2.6.3
